Time: 0 ms (100%), Space: 41.3 MB (8.08%) - LeetHub

+++ 0069-sqrtx/0069-sqrtx.java
@@ -1,5 +1,5 @@
 class Solution {
     public int mySqrt(int x) {
-        return (int)Math.sqrt((double)x);
+        return (int)Math.sqrt(x);
     }
 }
